Revert "[runtime] Fix potential overflow when using mono_msec_ticks"

Commit migrated from https://github.com/mono/mono/commit/c4ae0fa832464e3508aa84c3342bfbedebfb363f

+++ src/mono/mono/metadata/threads.c
@@ -3790,7 +3790,7 @@ mono_threads_abort_appdomain_threads (MonoDomain *domain, int timeout)
 #endif
 
 	abort_appdomain_data user_data;
-	gint64 start_time;
+	guint32 start_time;
 	int orig_timeout = timeout;
 	int i;
